requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
av==12.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
decord==0.6.0
# diffusers @ git+https://github.com/huggingface/diffusers.git@6f1042e36cd588a7b66498f45c3bb7085e4fa395
einops==0.8.1
exceptiongroup==1.3.1
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
Jinja2==3.1.6
kiwisolver==1.4.9
lightning-utilities==0.15.2
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.7.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
peft==0.17.0
pillow==11.3.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
# python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sentry-sdk==2.45.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.4
# torch==2.5.1+cu124
# torchaudio==2.5.1+cu124
torchmetrics==1.8.2
# torchvision==0.20.1+cu124
tqdm==4.67.1
transformers==4.48.1
triton==3.1.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wandb==0.20.1
wcwidth==0.2.14
yarl==1.22.0
zipp==3.23.0
